Add electron based "backend"
Summary: For non-Chrome environments (like RN) we use the Electron app standalone devtools, however it needs a backend to connect to. While previously this was baked into relay-runtime, in an effort to debug non-dev builds and reduce the relay modern runtime size, this adds a stand-alone backend that should be required into the application runtime before a Relay Environment is constructed.

Reviewed By: kassens

Differential Revision: D5803063

fbshipit-source-id: 8b2b1c183737d28462bb9b22692b4da005bbcd3e