Re-enable updates view in EnvironmentAgent
Summary:
The hooked EnvironmentAgent should emit events for store updates as a result of mutations to match existing devtools behavior.

As a bonus, this generalizes the view from "mutations" to "updates" which includes changes to the store as a result of queries and subscription payloads as well as from local updates unassociated with network activity.

Reviewed By: kassens

Differential Revision: D5818159

fbshipit-source-id: b1d20843848f0268afa5c6faca8ee643e9b98ae4