Dramatic speed up of updates view
Summary: Previously, every network or store update would result in two full copies of the relay record store, serializes both over the bridge, and performs a diff of every record, which can get very expensive for non-trivial apps. This replaces it with a version that uses updatedRecordIds to only copy those records which have changed as a result of the last operation, dramatically improving the speed of recording updates.

Reviewed By: kassens

Differential Revision: D5822771

fbshipit-source-id: 4020a4e13fe5db324fc1db645f3e88ad4fac5f59