Silence WebSocket failures
Summary:
When running Jest tests in RN in fbsource, relay-devtools warning that WebSocket is not defined takes a lot of space and attention:

{F111064900}

I have no idea how to rebuild compiled version of Relay checked into `xplat/js`. Please help. Or feel free to comandeer the revision and do the right thing.

Reviewed By: leebyron

Differential Revision: D5965057

fbshipit-source-id: 4864055dcddb4c82e888d152e6a307ed8a4b515a