remove dispatcher hack
Summary:
The combined reducer was post-processed to have the output of one reducer be
input into another one. This seemed pretty confusing and is now replaced by
consuming the same action in multiple reducers which IMHO is a totally fair way.

There's more to simplify with the navigation component, but that's in a
followup.

Reviewed By: jstejada

Differential Revision: D7464445

fbshipit-source-id: c460cc6b6dbf4c98dd00d8a31fc1afa12b11a279