Fix error in the network tab
Summary:
We have `execute.*` events in the dev tools for backward compatibility, and they are replaced with `network.*` events.
Now there are new `execute.*` events that are no longer network events, processing them causes the error, because the code is trying to read networkID or transactionID from `execute.*` events but these `id` no longer exist.

The fix is to remove all code related to `execute.*`, the backward compatibility will be broken, but I think it should be fine since old `execute.*` were replaced a while ago.

Reviewed By: alunyov

Differential Revision: D29347630

fbshipit-source-id: 07a33ec3bb4779548aa6749b06aa16aebef4544f