Setup logger in Relay Devtools
Summary:
Most code is a fork from React DevTools. A slight difference is `__ENABLE_LOGGER__` is set as a constant in webpackconfig.

This will load an iframe for logging in our devtools, and we will use the iframe to log devtools events.
`process.env.LOGGING_URL` is used to define the iframe endpoint, and it will be hidden from OSS.

Reviewed By: alunyov

Differential Revision: D32854824

fbshipit-source-id: 552b07d551d874ea2ccbe6760e1436fb9e4ad18f