Build script for internal bundle
Summary:
Adds a script to reduce the manual work involved to publish a version. The script automate bundling and `jf upload`, the rest of manual work involved is: bump the version in devtools manifest, and open a WWW ondemand to download the newly created bundle.

The `fb` folder isn't synced to github so it should be safe to keep fb only files. The `pem` file is commited directly to the codebase similar to what react devtools is doing. This should be safe because even people have access to `pem` file, they still need permission to publish to chrome store, so there is no way to affect public users.

Reviewed By: alunyov

Differential Revision: D32855293

fbshipit-source-id: 21b6896ed24704f1f3a8d786ffd93f6ddece3997