Optimize data passing from backend to frontend
Summary:
The devtools does a `JSON.parse(JSON.stringify(data))` on all Relay events to get around the limitation that all data passed through postMessages need to be serializable. But this is getting expensive as we add more information in Relay loggers.

This diff gets rid of the JSON calls, and instead preprocess data and filter out unused expensive information.

Reviewed By: andreqi

Differential Revision: D60936078

fbshipit-source-id: ba64679ee0df8db729dd2ec1d50e3c2a37c999be